Analysis: between-study heterogeneity and estimators in fMRI data.
Analysis where we gather real data and do an fMRI meta-analysis.
Studies are related to pain vs no pain.
Database is used to estimate the amount of between-study heterogeneity.
And then to plot different estimators that can be used to estimate
between-study heterogeneity when you work with standardized effect sizes.